util_dacfifo: Reduce logic on high fan-out dma_wren signal
The DMAC implementation guarantees that the expression `dma_valid &
dma_xfer_req` is always identical to just dma_valid.

When generating the util_dacfifo dma_wren signal the optimizer doesn't know
of this though and hence will route both signals into the LUT that drives
the write enable for the BRAMs.

Simplify the expression by removing dma_xfer_req from it. Considering this
can be a fairly high fan-out net and is typically the bottle-neck for the
util_dacfifo timing it helps to improve the timing.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>